requirements.txt
flask==3.1.*
psycopg2-binary==2.9.*
locust==2.32.*
